Add files via upload (Updated to 1.3.3-SNAPSHOT!)
+ Bug fixes (バグ修正)
(EN)
- Fixed an issue where using a part of the command causes an error when the party member is offline
- Fixed problem that language file overwritten at startup
(JP)
- パーティーメンバーがオフラインの時、一部のコマンドを使用するとエラーが発生する問題を修正
- 言語ファイルが起動時上書きされる問題を修正

+ Added support language (対応言語の追加)
(EN)
- Add arabic language
(JP)
- アラビア語の追加

+++ src/main/java/net/simplyrin/bungeeparties/utils/LanguageManager.java
@@ -51,14 +51,17 @@ public class LanguageManager {
 			languageFolder.mkdir();
 		}
 
-		String[] langs = { "english", "japanese", "chinese" };
+		String[] langs = { "english", "japanese", "chinese", "arabic" };
 		for(String lang : langs) {
-			try {
-				InputStream inputStream = LanguageManager.this.plugin.getResourceAsStream(lang + ".yml");
-				FileOutputStream outputStream = new FileOutputStream(new File(languageFolder, lang + ".yml"));
-				ByteStreams.copy(inputStream, outputStream);
-			} catch (Exception e) {
-				e.printStackTrace();
+			File languageFile = new File(languageFolder, lang + ".yml");
+			if(!languageFile.exists()) {
+				try {
+					InputStream inputStream = LanguageManager.this.plugin.getResourceAsStream(lang + ".yml");
+					FileOutputStream outputStream = new FileOutputStream(new File(languageFolder, lang + ".yml"));
+					ByteStreams.copy(inputStream, outputStream);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 			}
 		}
 	}
